requirements.txt
ai71==0.0.18
cohere==4.7.0
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
arize-phoenix==4.19.0
arize-phoenix-evals==0.14.1
asttokens==2.4.1
attrs==23.2.0
Authlib==1.3.1
backoff==2.2.1
blinker==1.8.2
boto3==1.34.153
botocore==1.34.153
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
Cython==0.29.37
datasets==2.14.7
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
dspy==0.1.5
dspy-ai==2.4.5
executing==2.0.1
fastapi==0.112.0
fastavro==1.9.5
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2023.10.0
googleapis-common-protos==1.63.2
graphql-core==3.2.3
greenlet==3.0.3
grpcio==1.65.2
grpcio-health-checking==1.65.2
grpcio-tools==1.65.2
h11==0.14.0
hdbscan==0.8.37
httpcore==1.0.5
httpx==0.25.2
httpx-sse==0.4.0
huggingface-hub==0.24.3
idna==3.7
importlib_metadata==8.0.0
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.43.0
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai==1.37.1
openinference-instrumentation==0.1.12
openinference-instrumentation-dspy==0.1.10
openinference-instrumentation-langchain==0.1.26
openinference-instrumentation-llama-index==2.2.1
openinference-instrumentation-openai==0.1.12
openinference-semantic-conventions==0.1.9
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
optuna==3.6.1
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pynndescent==0.5.13
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlean.py==3.45.1
stack-data==0.6.3
starlette==0.37.2
strawberry-graphql==0.236.0
structlog==24.4.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.19.1
tqdm==4.66.4
traitlets==5.14.3
types-requests==2.32.0.20240712
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
umap-learn==0.5.6
urllib3==2.2.2
uvicorn==0.30.5
validators==0.33.0
wcwidth==0.2.13
weaviate-client==4.7.1
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
PyMuPDF